fix(register): исправление замечаний после ревью
