+++ src/auth/auth.controller.ts
@@ -1,7 +1,7 @@
 import { Body, Controller, HttpCode, HttpStatus, Post } from '@nestjs/common';
-import { AuthService } from './auth.service';
 import { RegisterDto } from '../auth/dto/register-user.dto';
 import { Tokens } from '../auth/types';
+import { AuthService } from '../auth/auth.service';
 
 @Controller('auth')
 export class AuthController {
