+++ src/auth/auth.module.ts
@@ -4,25 +4,25 @@ import { AuthController } from './auth.controller';
 import { JwtRefreshStrategy } from './strategies/jwt-refresh.strategy';
 import { JwtAccessStrategy } from './strategies/jwt-access.strategy';
 import { PassportModule } from '@nestjs/passport'; // npm install --save-dev @types/passport-jwt
-import { ConfigModule } from '@nestjs/config';
-import { jwtConfig } from '../config/jwt.config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from '../entities/user.entity';
-import { RefreshToken } from '../entities/refreshToken.entity';
-import { UserRepository } from '../repository/register-user.repository';
+import { RefreshToken } from '../users/entities/refreshToken.entity';
+import { JwtService } from '@nestjs/jwt';
+import { ConfigService } from '@nestjs/config';
+import { UsersModule } from '../users/users.module';
 
 @Module({
   imports: [
+    UsersModule,
     PassportModule,
-    ConfigModule.forFeature(jwtConfig),
-    TypeOrmModule.forFeature([User, RefreshToken]),
+    TypeOrmModule.forFeature([RefreshToken]),
   ],
   controllers: [AuthController],
   providers: [
     AuthService,
     JwtRefreshStrategy,
     JwtAccessStrategy,
-    UserRepository,
+    JwtService,
+    ConfigService,
   ],
 })
 export class AuthModule {}
