+++ src/auth/auth.service.ts
@@ -1,44 +1,37 @@
-import { ConflictException, Injectable } from '@nestjs/common';
-import { RegisterDto } from 'src/auth/dto/register-user.dto';
-import { UserRepository } from '../repository/register-user.repository';
+import { Injectable } from '@nestjs/common';
 import { JwtConfig } from '../config/jwt.config';
-import { User } from '../entities/user.entity';
 import { JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
-import * as bcrypt from 'bcrypt';
+import { User } from '../users/entities/user.entity';
+import { RefreshToken } from '../users/entities/refreshToken.entity';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+import { UsersService } from '../users/users.service';
+import { RegisterDto } from '../auth/dto/register-user.dto';
 
 @Injectable()
 export class AuthService {
   constructor(
-    private readonly userRepository: UserRepository,
+    private readonly usersService: UsersService,
     private readonly jwtService: JwtService,
     private readonly configService: ConfigService,
+    @InjectRepository(RefreshToken)
+    private refreshTokensRepository: Repository<RefreshToken>,
   ) {}
 
   async register(registerDto: RegisterDto) {
-    const findUser = await this.userRepository.findUserByEmail(
-      registerDto.email,
-    );
-    if (findUser) {
-      throw new ConflictException(
-        `Пользователь с ${registerDto.email} уже существует!`,
-      );
-    }
+    const user = await this.usersService.register(registerDto);
+    const tokens = await this.generateTokens(user);
 
-    const hashedPassword = await bcrypt.hash(registerDto.password, 10);
-
-    const user = await this.userRepository.createUser(
-      registerDto,
-      hashedPassword,
-    );
-
-    const { accessToken, refreshToken } = await this.generateTokens(user);
-
-    return { accessToken, refreshToken };
+    return { ...tokens };
   }
 
-  private async generateTokens(user: User) {
-    const payload = { email: user.email, sub: user.id, role: user.userRole };
+  async generateTokens(user: User) {
+    const payload = { email: user.email, sub: user.id, role: user.role };
+
+    if (!payload.email || !payload.sub || !payload.role) {
+      throw new Error(`Отсутсвуют данные payload`);
+    }
 
     const jwtConfig = this.configService.get<JwtConfig>('JWT_CONFIG');
 
@@ -56,8 +49,26 @@ export class AuthService {
       secret: jwtConfig.refreshToken,
     });
 
-    await this.userRepository.createRefreshToken(user.id, refreshToken);
+    await this.createRefreshToken(user.id, refreshToken);
 
     return { accessToken, refreshToken };
+  }
+
+  async createRefreshToken(
+    userId: string,
+    refreshToken: string,
+  ): Promise<RefreshToken> {
+    const user = await this.usersService.findUserById(userId);
+
+    if (!user) {
+      throw new Error(`Пользователь с ${userId} не найден`);
+    }
+
+    const tokenEntity = this.refreshTokensRepository.create({
+      refreshToken,
+      user,
+    });
+
+    return await this.refreshTokensRepository.save(tokenEntity);
   }
 }
