--- src/auth/dto/refreshToken.dto.ts
@@ -1,6 +1,0 @@
-import { IsJWT } from 'class-validator';
-
-export class CreateRefreshTokenDto {
-  @IsJWT()
-  refreshToken: string;
-}
